customlearning(backend;build)/backend/requirements.txt
torch>=2.1.0
torchvision>=0.16.0
numpy>=1.25.0
pandas>=2.1.0
scikit-learn>=1.3.0
matplotlib>=3.8.0
Pillow>=10.0.0
